requirements.txt
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
Brotli==1.0.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
ImageHash==4.3.2
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.0
jupyter_server==2.15.0
jupyter_server_terminals==0.4.4
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.6
mistune==3.1.2
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numpy==2.2.4
opencv-python==4.11.0.86
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pillow==11.1.0
platformdirs==3.10.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
PyWavelets==1.8.0
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.13.1
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
win-inet-pton==1.1.0